1.27 commit (StatsStandard 다이얼로그 구현완료, SharedPreferences를 사용해 기록 측정중 인터넷 연결 끊길 시 기록 임시 저장 구현 중)

+++ app/src/main/java/com/example/runningeveryday/MeasureFragment.kt
@@ -276,7 +276,7 @@ class MeasureFragment : Fragment() {
 
 
         documentReference.collection(calendar.get(Calendar.DAY_OF_MONTH).toString())
-            .document(targetDistance.toString()).set(timeData as Map<String, Any>)
+            .document(targetDistance.toInt().toString()).set(timeData as Map<String, Any>)
 
 
         val top10Reference =
